deps: update to ngtcp2 0.8.1 and nghttp3 0.7.0
Refs: https://github.com/nodejs/node/pull/44619
PR-URL: https://github.com/nodejs/node/pull/44622

+++ deps/ngtcp2/nghttp3/lib/nghttp3_tnode.c
@@ -82,7 +82,7 @@ static uint64_t pq_get_first_cycle(nghttp3_pq *pq) {
 }
 
 int nghttp3_tnode_schedule(nghttp3_tnode *tnode, nghttp3_pq *pq,
-                           size_t nwrite) {
+                           uint64_t nwrite) {
   uint64_t penalty = nwrite / NGHTTP3_STREAM_MIN_WRITELEN;
 
   if (tnode->pe.index == NGHTTP3_PQ_BAD_INDEX) {
